Fixed warnings

+++ src/tune/tune.cpp
@@ -50,7 +50,7 @@ std::istream& operator>>(std::istream& stream, Tune& t) {
     
     for(int i = 0; i < t.getLaneCount(); i++) {
         NoteStream& stream = t.getLane(i).stream;
-        for(int j = 0; j < stream.getSetterSize(); j++) {
+        for(size_t j = 0; j < stream.getSetterSize(); j++) {
             SetterNote& note = stream.getSetterNote(j);
             int id = note.getId();
             if(note.getGen() != nullptr) continue;
@@ -152,7 +152,7 @@ double Tune::getSample(double srate) {
         if(!newPNotes.empty()) 
             newNotes = true;
         std::vector<SetterNote> newSNotes = l.stream.GetStartingSetterNotes(t);
-        for(int j = 0; j < newPNotes.size(); j++) {
+        for(size_t j = 0; j < newPNotes.size(); j++) {
             std::cout << newPNotes[j] << "    ";
             l.player.addNote(newPNotes[j]);
         }
